Replace patterns inside files

ATTENTION: Be careful, there is no coming back from the effects of this function.

replace_pattern() search each line of a file and replace a specific pattern when found.

TIP:

replace_pattern(
  dir = normalizePath(readClipboard(), "/", mustWork = FALSE),
  pattern = "pattern", replacement = "pattern"
)

replace_pattern(dir = utils::choose.dir(), pattern, replacement)

Arguments

dir

(optional) a string indicating the directory of the files. This function will look up just for files, directories will not be affected. (default:: utils::choose.dir()).

pattern

An character vector indicating the pattern to look for. The default interpretation is a regular expression. This parameter will be used on stringr::str_replace_all().

replacement

An character vector indicating the replacement value. This parameter will be used on stringr::str_replace_all().

Value

An invisible NULL. This function don't aim to return values.

Examples

if (FALSE) {
file_name <- tempfile(tmpdir = tempfile())
dir_name <- dirname(file_name)
dir.create(dir_name)
file.create(file_name)

con <- file(file_name)
open(con, open = "r+")

data <- c("01:00:56", "", "Test", "")

writeLines(data, con = con)
close(con)

remove_pattern(
  dir = dir_name, pattern = "([0-1][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]",
  replacement = ""
  )

con <- file(file_name)
open(con, open = "r+")
readLines(con)

close(con)
}